backend/requirements.txt
# requirements.txt
Flask
tensorflow
numpy
Pillow